Extract watcher service pools
